TODO5

+++ DemoCPP/btreepage.h
@@ -38,10 +38,10 @@ int binary_search(Container& container, int first, int last, ObjType &object)
 }
 
 template <typename Container, typename ObjType>
-void insert_at(Container& container, ObjType object, int pos)
+void insert_at(Container& container, ObjType object, size_t pos)
 {
         // TODO: #5 replace int, long by types such as size_t
-       int size = container.size();
+       size_t size = container.size();
        for(int i = size-2 ; i >= pos ; i--)
                container[i+1] = container[i];
        container[pos] =  object;
